refactor: Refactor overlay provider structure and relocate content overlay controller
