+++ packages/src/context/provider/content-overlay-controller.tsx
@@ -1,0 +1,55 @@
+import { type FC, useEffect, useRef } from 'react';
+
+type OverlayControllerProps = {
+  overlayId: string;
+  isOpen: boolean;
+  close: () => void;
+  unmount: () => void;
+};
+
+type OverlayAsyncControllerProps<T> = Omit<OverlayControllerProps, 'close'> & {
+  close: (param: T) => void;
+};
+
+export type OverlayControllerComponent = FC<OverlayControllerProps>;
+export type OverlayAsyncControllerComponent<T> = FC<OverlayAsyncControllerProps<T>>;
+
+type ContentOverlayControllerProps = {
+  isOpen: boolean;
+  current: string | null;
+  overlayId: string;
+  onMounted: () => void;
+  onCloseModal: () => void;
+  onExitModal: () => void;
+  controller: OverlayControllerComponent;
+};
+
+export function ContentOverlayController({
+  isOpen,
+  current,
+  overlayId,
+  onMounted,
+  onCloseModal,
+  onExitModal,
+  controller: Controller,
+}: ContentOverlayControllerProps) {
+  const prevCurrent = useRef(current);
+  const onMountedRef = useRef(onMounted);
+
+  /**
+   * @description Executes when closing and reopening an overlay without unmounting.
+   */
+  if (prevCurrent.current !== current) {
+    prevCurrent.current = current;
+
+    if (current === overlayId) {
+      onMountedRef.current();
+    }
+  }
+
+  useEffect(() => {
+    onMountedRef.current();
+  }, []);
+
+  return <Controller overlayId={overlayId} isOpen={isOpen} close={onCloseModal} unmount={onExitModal} />;
+}
