+++ packages/src/context/provider/use-sync-overlay-store.ts
@@ -1,5 +1,5 @@
 import { useSyncExternalStore } from 'use-sync-external-store/shim/index.js';
-import { type OverlayStore } from './store';
+import { type OverlayStore } from '../store';
 
 export function useSyncOverlayStore(overlayStore: OverlayStore) {
   const {
